Fixt the karma
Need to make a default value for karmathreshold in the Decisions.Choice class in case there is no karma condition. Making a really large negative karma threshold would be dumb in case there was a cyclic choice route that allowed you to farm karma to reach a massive negative value.